Move some code to `separate file` to `clarify` intent
