+++ Employee Information Management App/Employee Information Management App/Service/EmployeeLocalDataManager.swift
@@ -8,7 +8,6 @@
 import Foundation
 
 protocol EmployeeLocalDataManager {
-
     typealias AddEmployeeResult = Swift.Result<Employee, Error>
     typealias AddEmployeeResultCompletion = (AddEmployeeResult) -> Void
     func addEmployee(firstName: String, lastName: String, role: String, completion: @escaping AddEmployeeResultCompletion)
