+++ Employee Information Management App/Employee Information Management App/Service/LoginLocalDataManager.swift
@@ -14,10 +14,3 @@ protocol LoginLocalDataManager {
     func login(userName: String, password: String, completion: @escaping LoginResultCompletion)
     func hasSession() -> Bool
 }
-
-protocol LogoutLocalDataManager {
-    typealias LogoutResult = Swift.Result<(), Error>
-    typealias LogoutResultCompletion = (LogoutResult) -> Void
-
-    func logout(completion: @escaping LogoutResultCompletion)
-}
