+++ Employee Information Management App/Employee Information Management App/Service/LogoutLocalDataManager.swift
@@ -1,0 +1,15 @@
+//
+//  LogoutLocalDataManager.swift
+//  Employee Information Management App
+//
+//  Created by Jayco Bea on 11/15/22.
+//
+
+import Foundation
+
+protocol LogoutLocalDataManager {
+    typealias LogoutResult = Swift.Result<(), Error>
+    typealias LogoutResultCompletion = (LogoutResult) -> Void
+
+    func logout(completion: @escaping LogoutResultCompletion)
+}
